requirements.txt
git+https://github.com/openai/whisper.git
torch
torchvision
torchaudio
farm-haystack[all]
sentence-transformers
pandas
numpy
jupyter